BackEnd Refactor делегирования класса модели модуля рассписания
